requirements.txt
Python 3.12.5
pip 24.2
python-dotenv 1.0.1
google.generativeai 0.7.2